Move epsilon statement into its own class
