+++ boomerangPDS/src/main/java/boomerang/poi/CopyAccessPathChain.java
@@ -15,9 +15,9 @@
 package boomerang.poi;
 
 import boomerang.BackwardQuery;
+import boomerang.scope.ControlFlowGraph;
 import boomerang.scope.ControlFlowGraph.Edge;
 import boomerang.scope.Field;
-import boomerang.scope.Statement;
 import boomerang.scope.Val;
 import boomerang.scope.ValCollection;
 import boomerang.solver.BackwardBoomerangSolver;
@@ -104,9 +104,7 @@ public class CopyAccessPathChain<W extends Weight> {
       }
       INode<Node<Edge, Val>> targetState =
           backwardSolver.generateFieldState(
-              new SingleNode<>(
-                  new Node<>(
-                      new Edge(Statement.epsilon(), Statement.epsilon()), ValCollection.zero())),
+              new SingleNode<>(new Node<>(ControlFlowGraph.Edge.epsilon(), ValCollection.zero())),
               t.getLabel());
       Transition<Field, INode<Node<Edge, Val>>> insert =
           new Transition<>(stateInBwSolver, t.getLabel(), targetState);
